+++ libraries/test_session_current/src/main/java/androidx/media3/session/MediaControllerCompatProviderService.java
@@ -156,6 +156,12 @@ public class MediaControllerCompatProviderService extends Service {
       controller.sendCommand(command, params, cb);
     }
 
+    @Override
+    public String getPackageName(String controllerId) throws RemoteException {
+      MediaControllerCompat controller = mediaControllerCompatMap.get(controllerId);
+      return controller.getPackageName();
+    }
+
     ////////////////////////////////////////////////////////////////////////////////
     // MediaControllerCompat.TransportControls methods
     ////////////////////////////////////////////////////////////////////////////////